refactor cli into drive command
